split routes from server.js
